resolved reviews and conflicts

+++ ExpenseTracker/Record/Expense.cs
@@ -8,7 +8,8 @@
         /// <summary>
         /// Amount of the record
         /// </summary>
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
+
         /// <summary>
         /// Category of the record
         /// </summary>
@@ -19,7 +20,7 @@
         /// </summary>
         /// <param name="amount">Amount of the expense</param>
         /// <param name="category">Category of the expense</param>
-        public Expense(int amount, string category)
+        public Expense(decimal amount, string category)
         {
             Amount = amount;
             Category = category;
